% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eagle2.R
\name{detect_homozygotes_w_replicatess}
\alias{detect_homozygotes_w_replicatess}
\title{Detect potential homozygote (or imprinted) SNP-individual pairs and zero them out.}
\usage{
detect_homozygotes_w_replicatess(a, nh, concShape = 1.0001,
  concRate = 1e-04, errorRate = 0.01, posterior_theshold = 0.95,
  verbose = T)
}
\arguments{
\item{a}{Numerator counts [n x T x K] where n are individuals, T are timepoints, K are SNPs}

\item{nh}{denominator counts [n x T x K]}

\item{concShape}{Shape of prior on concentration}

\item{concRate}{Rate of prior on concentration}

\item{errorRate}{Assumed probability of erroneously observing a read with the alternative allele, due to sequencing or alignment error}

\item{posterior_threshold}{If posterior probability of being a het is below this the individual/SNP pair will be ignored.}
}
\value{
Filtered a and nh matrices, and which SNPs were kept.
}
\description{
Detect potential homozygote (or imprinted) SNP-individual pairs and zero them out.
}
