% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eagle2.R
\name{eagle2}
\alias{eagle2}
\title{Beta binomial GLM with flips. Prior on concentration parameter is Gamma(concShape,concRate)}
\usage{
eagle2(ys, ns, concShape = 1.0001, concRate = 1e-04, hessian = T,
  ...)
}
\arguments{
\item{ys}{numerator counts [n x T x K] where n are individuals, T are timepoints, K are SNPs}

\item{ns}{denominator counts [n x T x K]}

\item{concShape}{Shape of prior on concentration}

\item{concRate}{Rate of prior on concentration}
}
\value{
List with likelihood ratio, p-value and fits
}
\description{
Beta binomial GLM with flips. Prior on concentration parameter is Gamma(concShape,concRate)
}
