% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eagle2.R
\name{get_coefs}
\alias{get_coefs}
\title{Function to extract coefficients, standard errors, and Wald p-values.}
\usage{
get_coefs(fit)
}
\arguments{
\item{fit}{Fitted object returned from rstan::optimizing. Must have $par$beta and $hessian available}
}
\value{
Data.frame with coefs, standard errors, Wald p-values
}
\description{
Function to extract coefficients, standard errors, and Wald p-values.
}
