% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eagle2.R
\name{robustSolve}
\alias{robustSolve}
\title{Robust inverse of PSD matrix using SVD}
\usage{
robustSolve(si, eigenThreshold = 0.01)
}
\arguments{
\item{si}{Matrix to invert}

\item{eigenThreshold}{Threshold eigenvalues smaller than this}
}
\value{
Inverted matrix
}
\description{
Robust inverse of PSD matrix using SVD
}
