% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svem.R
\name{svem}
\alias{svem}
\title{Stochastic variational EM.}
\usage{
svem(grad_log_prob, to_optim, init = NULL, plot.elbo = F,
  log_prob = NULL, master_stepsize = 1, ...)
}
\arguments{
\item{grad_log_prob}{Function which takes unconstrained parameters and returns the gradient of the log joint distribution.}

\item{to_optim}{Logical vector denoting which parameters should be optimized over rather than integrated over using SVI.}

\item{init}{Initialization.}

\item{plot.elbo}{Whether to plot the progress in term of the ELBO (Evidence Lower Bound)}

\item{log_prob}{(optional) Function returning log joint distribution.}
}
\value{
List with
\item{m}{Mean of approximate posterior for each parameter (or point estimate)}
\item{s}{Standard deviation of approximate posterior for each param, or 0 or optimized params}
\item{elbo_func}{Function to calculate (noisy) estimate of ELBO using a single sample from rnorm(sum(!to_optim)).}
\item{elbo_progress}{Vector of estimated ELBO with algorithm iteration.}
}
\description{
Uses Stochastic Variational Inference (SVI) over parameters where to_optim==F and regular Stochastic Gradient (well Adagrad) over parameters for which to_optim==T.
}
